requirements.txt
Django==1.9.1
djangorestframework==3.4.4
six==1.10.0
psycopg2==2.6.2
django-extensions==1.6.7
Pillow==3.4.2
